NLP/requirements.txt
flask
flask-cors
requests
pyttsx3
gunicorn
python-dotenv
nltk
spacy 